added readme file, ready for styling and ux

+++ src/App.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import { Button } from '@material-ui/core'
-import logo from './logo.svg';
 import { BrowserRouter as Router, Route, Link, BrowserRouter } from 'react-router-dom'
 import './App.css';
 import {
@@ -14,39 +13,29 @@ import {
   components,
 } from 'react-big-calendar'
 import moment from 'moment'
-
-// import Navbarcomp from './components/actions/navbarcomp'
-// import Clients from './components/clients/clients';
-// import Modal from 'react-bootstrap/Modal'
 import Popup from './components/actions/popa';
-// import Analytics from './components/analytics/analytics';
-// import { LineChart, Line } from 'recharts';
 import axios from 'axios';
-import { withTranslation } from 'react-i18next';
-import { MDCRipple } from '@material/ripple';
 import { FloatingActionButton } from "material-ui";
 import { Add } from 'material-ui-icons';
 import PropTypes from 'prop-types'
 import Editable from './components/actions/editable';
 
-let dateFormat = PropTypes.any;
+// let dateFormat = PropTypes.any;
 let dateRangeFormat = PropTypes.func;
 
-const timeRangeStartFormat = ({ start }, culture, local) =>
-  `${local.format(start, 't', culture)} — `
+// const timeRangeStartFormat = ({ start }, culture, local) =>
+//   `${local.format(start, 't', culture)} — `
 
-const timeRangeEndFormat = ({ end }, culture, local) => ` — ${local.format(end, 't', culture)}`
+// const timeRangeEndFormat = ({ end }, culture, local) => ` — ${local.format(end, 't', culture)}`
 
-const timeRangeFormat = ({ start, end }, culture, local) =>
-  `${local.format(start, 't', culture)} — ${local.format(end, 't', culture)}`
+// const timeRangeFormat = ({ start, end }, culture, local) =>
+//   `${local.format(start, 't', culture)} — ${local.format(end, 't', culture)}`
 
 let localizer = momentLocalizer(moment)
-// let allViews = Object.keys({Views}).map(k => console.log({Views}))
 
 class App extends Component {
   constructor() {
     super();
-    // this.getDatafromDB()
     this.state = {
       data: [],
       showPopup: false,
@@ -55,29 +44,29 @@ class App extends Component {
 
   }
   async componentDidMount() {
-    // console.log(this.state.data)
     await this.getDatafromDB()
   }
 
   async getDatafromDB() {
     let data = await axios.get('http://localhost:4328/events')
-    console.log(data.data)
     return this.setState({ data: data.data, showPopup: false })
   }
 
-  async delEventfromDB(eventid) {
-    let _id = eventid
-    let data = await axios.delete('http://localhost:4328/delevent', {
-      params: { _id }
-    })
-    console.log(data.data)
-    return this.setState({ data: data.data, showPopup: false })
+  async delEventfromDB(_id) {
+    if (window.confirm("Delete this reminder ? ")) {
+      let data = await axios.delete('http://localhost:4328/delevent', {
+        params: { _id }
+      })
+      alert('Reminder Deleted!')
+      return this.setState({ data: data.data, editmode: false })
+    } else {
+      return this.setState({ editmode: false })
+    }
   }
 
   async delDatafromDB(eventid) {
     if (window.confirm("Are you sure you want to empty your reminders .")) {
-      let data = await axios.delete('http://localhost:4328/empty', {
-      })
+      let data = await axios.delete('http://localhost:4328/empty', {})
       alert('Deleted!.')
       return this.setState({ data: [], showPopup: false })
     } else {
@@ -86,10 +75,7 @@ class App extends Component {
   }
 
   togglePopup = (event) => {
-    // console.log(e.target.id)
-    console.log(event)
     this.setState({ showPopup: true })
-    // this.props.togglePopup()
   }
 
   closePopup = () => {
@@ -97,6 +83,7 @@ class App extends Component {
       showPopup: null
     })
   }
+
   closeEdit = () => {
     return this.setState({
       editmode: false
@@ -106,7 +93,6 @@ class App extends Component {
   selecEvent = (event) => {
     let editmode = { ...this.state.editmode }
     editmode = JSON.stringify(event)
-    console.log(editmode)
     this.setState({ editmode })
     return
   }
@@ -117,10 +103,6 @@ class App extends Component {
 
   pushData = async (title, start, end, allDay, city, color) => {
     let data = {}
-    console.log(end)
-    // let data = [...this.state.data]
-    // data.push({ name, country, owner })
-    // let nclient = {name, country, owner}
     await axios.post('http://localhost:4328/pevent', {
       data: { title, start, end, allDay, city, color }
     })
@@ -129,51 +111,36 @@ class App extends Component {
 
   updateEvent = async (id, title, start, end, allDay, city, color) => {
     let data = {}
-    console.log(end)
-    // let data = [...this.state.data]
-    // data.push({ name, country, owner })
-    // let nclient = {name, country, owner}
     await axios.put('http://localhost:4328/upevent', {
       data: { id, title, start, end, allDay, city, color }
     })
-    console.log(data)
     return this.getDatafromDB()
   }
 
   render() {
-    const { t, i18n } = this.props;
-    const changeLanguage = lng => {
-      i18n.changeLanguage(lng);
-      console.log({ Views })
-    };
     let formats = {
-      // dateFormat: 'dd',
-      // dayFormat: (date, format,  localizer) =>
-      //   localizer.format(date, 'DDD', culture),
       eventTimeRangeFormat: dateRangeFormat,
       eventTimeRangeStartFormat: dateRangeFormat,
       eventTimeRangeEndFormat: dateRangeFormat
-      // timeGutterFormat: 't'
     }
 
     return <div className="App">
       <BrowserRouter>
         <Route path="/" exact render={() =>
           <div className="calendar-container">
-            <Button variant="outlined" onClick={this.togglePopup} name="showPopup" id="showPopup">Add Reminder</Button>
-            <Button variant="contained" onClick={this.delDatafromDB}
-              position='end' color="secondary" name="delete all" id="empty">Empty ALL</Button>
-            {this.state.editmode ? <Editable selectEvent={this.state.editmode} closeEdit={this.closeEdit} updateEvent={this.updateEvent} /> :
-              <></>}
+            <Button variant="outlined" onClick={this.togglePopup} name="showPopup" id="showPopup"
+            >Add Reminder</Button>
+            <Button variant="contained" onClick={this.delDatafromDB} position='end' color="secondary" name="delete all" id="empty"
+            >Empty ALL</Button>
+
+            {this.state.editmode ? <Editable delEventfromDB={this.delEventfromDB}
+              selectEvent={this.state.editmode} closeEdit={this.closeEdit} updateEvent={this.updateEvent} /> : <></>}
+
             {this.state.showPopup ? <Popup closePopup={this.closePopup} pushData={this.pushData} /> :
               <Calendar
                 formats={formats}
                 localizer={localizer}
                 events={this.state.data}
-                // startAccessor="start"
-                // endAccessor="end"
-                // getNow={this.getNow}
-                culture="ES"
                 style={{ height: 500 }}
                 onSelectEvent={this.selecEvent}
                 views={['month']}
@@ -185,4 +152,4 @@ class App extends Component {
     </div>
   }
 }
-export default withTranslation('translation')(App);
+export default App;
